add profile data to alerts. tweak background hosted character queue dropping a bunch if Census was being wacky.

+++ Code/Commands/AlertCommand.cs
@@ -1,9 +1,12 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using watchtower.Commands;
+using watchtower.Models.Alert;
+using watchtower.Models.Census;
 using watchtower.Services.Db;
 using watchtower.Services.Repositories;
 
@@ -13,19 +16,36 @@ namespace watchtower.Code.Commands {
     public class AlertCommand {
 
         private readonly ILogger<AlertCommand> _Logger;
+        private readonly AlertRepository _AlertRepository;
         private readonly AlertPlayerDataRepository _DataRepository;
-        private readonly AlertParticipantDataDbStore _DataDb;
+        private readonly AlertPlayerDataDbStore _DataDb;
+        private readonly AlertPlayerProfileDataDbStore _ProfileDataDb;
 
         public AlertCommand(IServiceProvider services) {
             _Logger = services.GetRequiredService<ILogger<AlertCommand>>();
+            _AlertRepository = services.GetRequiredService<AlertRepository>();
             _DataRepository = services.GetRequiredService<AlertPlayerDataRepository>();
-            _DataDb = services.GetRequiredService<AlertParticipantDataDbStore>();
+            _DataDb = services.GetRequiredService<AlertPlayerDataDbStore>();
+            _ProfileDataDb = services.GetRequiredService<AlertPlayerProfileDataDbStore>();
         }
 
         public async Task Rebuild(long alertID) {
+            PsAlert? alert = await _AlertRepository.GetByID(alertID);
+            if (alert == null) {
+                _Logger.LogWarning($"Alert {alertID} doesn't exist");
+                return;
+            }
+
             _Logger.LogInformation($"Rebuilding participant data for alert {alertID}...");
+
+            await _ProfileDataDb.DeleteByAlertID(alertID);
             await _DataDb.DeleteByAlertID(alertID);
-            await _DataRepository.GetByAlert(alertID, CancellationToken.None);
+
+            List<AlertPlayerDataEntry> parts = await _DataRepository.GetByAlert(alertID, CancellationToken.None);
+            alert.Participants = parts.Count;
+
+            await _AlertRepository.UpdateByID(alertID, alert);
+
             _Logger.LogInformation($"Done rebuilding participant data for alert {alertID}");
         }
